+++ frontend/src/components/AnchorList.vue
@@ -489,7 +489,7 @@ import RankComparison from '@/components/RankComparison.vue'
 import MonthSelector from '@/components/MonthSelector.vue'
 import { getMonthRange } from '@/utils/monthUtils'
 import { provideGlobalCardState } from '@/composables/useGlobalCardState'
-import { getAvatar, getAvatarByUid, preloadAllAvatars, scaleAvatar, getAvatarSync, loadAvatarAndUpdate } from '@/utils/avatarCache'
+import { fetchAllAvatars, getAvatar, getAvatarByUid, preloadAllAvatars, scaleAvatar, getAvatarSync, loadAvatarAndUpdate } from '@/utils/avatarCache'
 
 Chart.register(...registerables)
 
@@ -823,6 +823,9 @@ export default {
       try {
         loading.value = true
         error.value = null
+
+        fetchAllAvatars()
+
         let response;
         const currentMonth = route.query.month || null;
         if (currentMonth) {